Refactor SharedDisposal
